requirements.txt
# Aggregated requirements for Railway deployment.
# Covers backend, ingestion, news-poller, and llm-synthesizer.

# Kafka (kafka-python-ng is a maintained fork; all services use it)
kafka-python-ng>=2.2.3

# Backend API
fastapi==0.111.0
uvicorn==0.30.1
sse-starlette==2.1.0
python-dotenv==1.0.0

# Ingestion
websockets==12.0
playwright>=1.40.0
playwright-stealth
certifi

# News / Market / PortWatch / Prediction pollers
feedparser==6.0.11
requests>=2.31.0

# LLM Synthesizer
anthropic>=0.34.0
